+++ strbuf.h
@@ -12,6 +12,6 @@
 
 char *fgetl(FILE *); // gets a line of string data; returns a malloc-like pointer
 char *slurp(FILE *); // gets an entire file of string data; returns a malloc-like pointer
-void init_char(char **buf, int *l, int *i); // initialises a string buffer in heap.  *buf becomes a malloc-like pointer
-void append_char(char **buf, int *l, int *i, char c); // adds a character to a string buffer in heap (and realloc()s if needed)
-void append_str(char **buf, int *l, int *i, const char *str); // adds a string to a string buffer in heap (and realloc()s if needed)
+void init_char(char **buf, size_t *l, size_t *i); // initialises a string buffer in heap.  *buf becomes a malloc-like pointer
+void append_char(char **buf, size_t *l, size_t *i, char c); // adds a character to a string buffer in heap (and realloc()s if needed)
+void append_str(char **buf, size_t*l, size_t *i, const char *str); // adds a string to a string buffer in heap (and realloc()s if needed)
